+++ srcs/reset.c
@@ -6,7 +6,7 @@
 /*   By: sakitaha <sakitaha@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/20 16:35:23 by sakitaha          #+#    #+#             */
-/*   Updated: 2024/02/22 21:28:24 by sakitaha         ###   ########.fr       */
+/*   Updated: 2024/03/14 13:33:45 by sakitaha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,11 +18,11 @@
  */
 static void	reset_scale(t_fdf *fdf)
 {
-	float	x_len;
-	float	y_len;
+	double	x_len;
+	double	y_len;
 
-	x_len = (float)WIN_WIDTH / fdf->max_x / 2;
-	y_len = (float)WIN_HEIGHT / fdf->max_y / 2;
+	x_len = (double)WIN_WIDTH / fdf->max_x / 2;
+	y_len = (double)WIN_HEIGHT / fdf->max_y / 2;
 	fdf->scale = fmin(x_len, y_len);
 	if (fdf->scale > 3)
 	{
@@ -36,8 +36,8 @@ static void	reset_scale(t_fdf *fdf)
  */
 static void	reset_move(t_fdf *fdf)
 {
-	fdf->x_move = ((float)WIN_WIDTH - (fdf->max_x - 1) * fdf->scale) / 1.6;
-	fdf->y_move = ((float)WIN_HEIGHT - (fdf->max_y - 1) * fdf->scale) / 2;
+	fdf->x_move = ((double)WIN_WIDTH - (fdf->max_x - 1) * fdf->scale) / 1.6;
+	fdf->y_move = ((double)WIN_HEIGHT - (fdf->max_y - 1) * fdf->scale) / 2;
 }
 
 /**
